Add files via upload
Company which has 200 workers and there are consultants who connects through their own devices and network from different places. Remote users who are from different location should get connected server throughout a public server as remote users there are developer from different locations whom also wanted to get connected to the server through their laptop and mobile phones. Stake holders who are located in UK and US & they can only login through the cloud and get updated about the company. 
The company stores its data to the main server and then sends a backup to the cloud which is located in a different location such in a foreign country. This is what the scenario says and according to the problem we’ve implemented solution for this. 
 
Existing system Problems 
 
 Unable to defend the server from harmful threats. Authentication policy is unavailable  Network is unable to protect backup, control and manage.  No centralized OS and a directory control database  No proper Networking plan  Many multiple client OS versions can run in remote PCs.